Добавлены доменные сущности

+++ migrations/struct/20220719191934_create_outlets_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE outlets (
     id TEXT PRIMARY KEY,
     station_id TEXT NOT NULL REFERENCES stations (id) ON DELETE CASCADE,
     connector INTEGER,
-    kilowatts INTEGER,
+    kilowatts FLOAT,
     power INTEGER
 );
